Update: Attributes with sub-attributes will not be returned anymore by projection
cc @leodido
The result of projection is usually converted to a map prior to be passed to the database and there's no way to make map's items order predictable, thus path to a complex attribute and paths to its sub-attributes could be returned in any order.
Within mongo, different projection's items ordering will produce different query results (in particular when the parent attribute is listed after its sub-attributes), so the final behavior is unpredictable because map's items ordering could be different each time.
Currently, to mitigate the "parent and children ordering" issue we decided to return only leaf-attributes that's not affect the final projection.
Also, with this strategy, it's not possibile to exclude all sub-attributes by just excluding their parent attribute (because it's always ignored) so it makes impossibile to implement the behavior described by #51